+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+# Kegg2bipartitegraph v0.2.0 (2024-12-03)
+
+## Add
+
+* `scope` subcommand to compute scope using graphml files and seed files.
+* Script (`seeds.py`) to create several seed files used in the [article of Webber-Zendrera et al. (2021)](https://doi.org/10.1038/s41598-021-91486-8) when creating reference model.
+* Warning to k2bg organism if there is a difference between the version of KEGG in the package and the version of KEGG online.
+
+## Fix
+
+* Issue with GitHub Actions.
+
+## Modify
+
+* Update reference model to KEGG 112.
+* Reference model now contains reaction to remove (either glycan or ubiquitous), they are removed when creating specific models.
+
+## Remove
+
+* Data folder containing seed files as they are now inside the package.
+
 # Kegg2bipartitegraph v0.1.2 (2024-11-26)
 
 ## Fix